chore: move files gateway service to dsn
